He UI lo ren 7 khung (ep do tim + ep Hoang Kim mon phai): hoi sinh box giao dien + vong phan bien 3 huong
Nen (port tu CLIENT VLTK, khong phai may chu Linux): KItemCompound.cpp/.h moi -
KFoundryResDemand doc settings\item\foundryresdemand.ini, 9 ham kiem theo
COMPOUND_TYPE, chan so o (s_anSoO), chan do dang xep chong, nhom khoa thay the.

Hoi sinh box 7 khung (bi cat day tu 2021):
- May chu: 6 phong do pos_compone..pos_enchase song lai theo khuon pos_tremble
  (6 mang + 18 ham Check/Add/Un, mo khoi thu hoi do); nut bam qua UiCommandScript
  case 7 voi danh sach trang 6 callback LR_UI_*; GetIdItem case 4..9 doc phong.
- Client: 6 enum GDI_COMPONE..GDI_ENCHASE + GetGameData + SWITCH_OBJECT + notify;
  6 nut gui GOI_ADD_UI_CMD_SCRIPT type 7.
- Sua 2 loi co san tu 2021: khung 3 SetContainerId chep nham UOC_COMPTWO;
  KUiForge::LoadScheme lap _ITEM_COMP_COUNT(3) tren mang 2 phan tu (tran that).

Vong phan bien 3 huong (mat do / an ninh - truc loi / client ve - dong bo) - 20 mieng va:
- DUP nguyen lieu xep chong: KItemCompound.cpp:918-935 tinh gia tri NHAN ca chong
  (dich nguoc Linux 0x08153F21) nhung RemoveItemByIndex 1 doi ben JX1 chi tru 1 don vi.
  Ban VLTK goc goi 1 doi vi engine cua ho xoa ca chong. Chi dinh 18 muc ConsumeItem_*
  (Stackable=1) vi KhopTietDoan:429 chan chong o moi o khac. Vá: truyen them
  GetItemStackCount(idx) trong compound_header.lua + ore_upgrade.lua.
- Khong huy do dang KHOA: them chot bo qua trong vong loc nguyen lieu.
- The Kham nam TAC hoan toan: nDetail=0 cung -> luon ReturnInfo 16, lenh gui khong
  bao gio toi. Doi sang -1 de roi default (may chu kiem that qua KFoundryResDemand).
- Cu bam DAU khong gui + treo anh hieu ung vinh vien o ca 6 pad: PlayEffect() rong
  return 1 lam Breathe ket, nhanh tu-gui unreachable. Bo nhanh "nap" -> 1 bam = 1 ep.
- Dong cua so chi thu hoi do cua the dang mo -> thu hoi ca 6 phong.
- Cua so bi bo quen khoi UiCloseWndsInGame: ESC khong dong duoc, giao dich/danh cuoc
  chong cua so, thoat game khong destroy -> o do MA xuyen phien. Them vao ca 3 cho.
- KUiCompoundThree::LoadScheme tra nham CtrlCompTwoItemMap; KUiDistill case 2 thieu break.

HOAN CO CHU Y: mieng dat m_dwCompoundItemId=0 tren duong c2s_recoveritem (giu trong
vA_server_loren.py) - ca 6 pad gui GOI_RECOVER_ITEM tu nut "Huy bo" ma KHONG dong cua so
nen reset se khoa cong m_dwCompoundItemId>0, moi cu bam ep sau im lang. Dot sau lam
bang callback dong cua so goi EndCompoundItem.

Build sach ca ba (0 loi), da dat len may chu, bam khop, byte cao TCVN3/GBK khong doi,
FFFD=0. CoreShell.cpp/.h + KCore.cpp co kem vai dong he kinh mach cua phien song song
(cung tep, khong tach duoc ma van build).

Chi tiet + viec con mo (the Kham nam dang muon bo cuc the Trich lay; the Atlas/HKMP
chua co lop cua so; g_ComItem menu tro the hien rong): BANGIAO_UILOREN_2708.md muc 0-BIS.

+++ Sources/Core/Src/KProtocol.h
@@ -2179,7 +2179,12 @@ typedef struct
 	BYTE	m_btSeries;			// ÎïÆ·µÄÎåÐÐ
 	BYTE	m_btLevel;			// ÎïÆ·µÄµÈ¼¶
 	BYTE	m_btLuck;			// MF
-	BYTE	m_btMagicLevel[MAX_ITEM_MAGICLEVEL];	// Éú³É²ÎÊý
+	// [LOREN] Ban Linux truyen truong nay bang int 32-bit: trong cung ham
+	// dong goi 0x081F9430, nLuck ra opcode 88 (MOV r/m8) con sau o cap
+	// thuoc tinh ra 89 (MOV r/m32), stride 4. Quet ca doan ma: moi truy
+	// cap vao +0x1E0..+0x200 deu rong 4 byte, khong luot nao 1 hay 2 byte.
+	// Bat buoc phai vay: magicattriblevel_index.txt trai kin 1..480.
+	int		m_btMagicLevel[MAX_ITEM_MAGICLEVEL];	// Éú³É²ÎÊý
 	WORD	m_wVersion;			// ×°±¸°æ±¾
 	DWORD	m_dwRandomSeed;		// Ëæ»úÖÖ×Ó
 	BYTE	m_bPoint;
